New Commands and Graph Fixes
2 new commands: listmaps and currentmap.
Updated the gamestats module to include Knights and Paladins in the Troop Summary.
Changed the Time column of each dataframe to be a datetime.time() object instead of just datetime (should fix the x-axis just saying the date over and over again.

Changed all path and directory references to use os.path objects, and based of global variable to switch between relative and absolute with ease.